Revert "rename and move BiomeType enum"
This reverts commit 0b5595a7aef00a18a3a0de7a0dd0abc83b1c9524.

+++ src/main/java/net/rom/exoplanets/astronomy/yzcetisystem/c/WorldProviderYzCetiC.java
@@ -62,12 +62,12 @@ public class WorldProviderYzCetiC extends WorldProviderExoPlanet {
 	@Override
 	public float getPlanetTemp() {
 		ExoPlanet planet = this.getPlanet();
-		float planetTemp = (float) planet.getPlanetTemp();
+		float planetTemp = planet.getPlanetTemp();
 
 		if (this.isDaytime()) {
 			planetTemp *= 4.5F;
 		} else {
-			planetTemp = (float) planet.getPlanetTemp();
+			planetTemp = planet.getPlanetTemp();
 		}
 		return planetTemp;
 	}
